+++ src/Constants.php
@@ -30,4 +30,15 @@ class Constants
     public const ELOQUENT_CREATED_AT = 'created_at';
     public const ELOQUENT_UPDATED_AT = 'updated_at';
     public const CHARACTERS = ['-', '.', ',', ';', ':', ' ', '?', '\'', '"', '#', '%', '&', '/', '\\', '(', ')'];
+    public const PHPDOC_HEADER = [
+        'This file has been auto-generated.',
+        'Update settings in config file for table',
+        'and run "make:models %s %s".',
+        ''
+    ];
+    public const ATTRIBUTES_FILLABLE = 'The attributes that are mass assignable.';
+    public const ATTRIBUTES_GUARDED = 'The attributes that aren\'t mass assignable.';
+    public const ATTRIBUTES_HIDDEN = 'The attributes that should be hidden for arrays.';
+    public const ATTRIBUTES_CASTS = 'The attributes that should be cast to native types.';
+    public const ATTRIBUTES_ACCESSORS = 'The accessors to append to the model\'s array form.';
 }