+++ tests/Builders/PhpDocBuilderTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Tests\CoRex\Laravel\Model\Builders;
 
 use CoRex\Laravel\Model\Builders\PhpDocBuilder;
+use CoRex\Laravel\Model\Constants;
 use Tests\CoRex\Laravel\Model\TestBase;
 
 class PhpDocBuilderTest extends TestBase
@@ -16,9 +17,14 @@ class PhpDocBuilderTest extends TestBase
     {
         $lines = $this->createBuilder(PhpDocBuilder::class)->build();
 
-        $this->assertSame(
+        $phpdocLines = ['/**'];
+        foreach (Constants::PHPDOC_HEADER as $line) {
+            $phpdocLines[] = ' * ' . sprintf($line, $this->connection, $this->table);
+        }
+
+        $phpdocLines = array_merge(
+            $phpdocLines,
             [
-                '/**',
                 ' * @property int $id',
                 ' * @property-read string $code',
                 ' * @property int $number',
@@ -27,8 +33,9 @@ class PhpDocBuilderTest extends TestBase
                 ' * @property string $created_at_test',
                 ' * @property string $updated_at_test',
                 ' */',
-            ],
-            $lines
+            ]
         );
+
+        $this->assertSame($phpdocLines, $lines);
     }
 }