+++ tests/migrations/2020_06_26_071424_create_ltest_table.php
@@ -10,14 +10,12 @@ class CreateLtestTable extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create(
             'ltest',
-            function (Blueprint $table) {
+            function (Blueprint $table): void {
                 $table->id();
                 $table->string('code', 50);
                 $table->string('value', 50);
@@ -28,10 +26,8 @@ class CreateLtestTable extends Migration
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::drop('ltest');
     }