+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Added support for showing destination of generated models (-d|--destination).
 - Added support for showing content of generated models (-c|--console).
 - Added support for replacing builders.
+- Added support for hidden attributes.
+- Added support for castable attributes.
+- Added support for appendable accessors.
 
 ### Removed
 - Removed option to specify guarded fields on commands, since it is possible to specify in config file.